// chpqa-app/src/app/api-services/chpqa-api/generated/model/requestUpdateSubmissionPowerOutput.ts
/**
 * WebApi
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { RequestPowerOutput } from './requestPowerOutput';


export interface RequestUpdateSubmissionPowerOutput { 
    idSubmission: string;
    totalPowerGenerated: number;
    totalPowerExported?: number | null;
    totalPowerImported?: number | null;
    powerOutputList: Array<RequestPowerOutput>;
}